requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
django-debug-toolbar==2.2
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.9
mysqlclient==1.4.6
Pillow==7.1.2
pysolr==3.9.0
python-memcached==1.59
python3-memcached==1.51
pytz==2019.3
qiniu==7.2.8
requests==2.23.0
shortuuid==1.0.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
Whoosh==2.7.4
